backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-dotenv==1.0.0
langchain==0.0.312
langchain-openai==0.0.2
langchain-core==0.1.1
faiss-cpu==1.7.4
pandas==2.1.0
numpy==1.24.3
openai==1.3.3
pydantic==1.10.8
python-multipart==0.0.6
typing-extensions==4.8.0 